+++ src/models/SettingsModel.php
@@ -9,7 +9,6 @@ use Craft;
 use craft\base\Model;
 use craft\behaviors\EnvAttributeParserBehavior;
 use putyourlightson\blitz\drivers\purgers\CloudflarePurger;
-use putyourlightson\blitz\drivers\purgers\KeyCdnPurger;
 use putyourlightson\blitz\drivers\storage\FileStorage;
 use putyourlightson\blitz\drivers\purgers\DummyPurger;
 
@@ -61,10 +60,7 @@ class SettingsModel extends Model
     /**
      * @var array
      */
-    public $cachePurgerTypes = [
-        CloudflarePurger::class,
-        KeyCdnPurger::class,
-    ];
+    public $cachePurgerTypes = [CloudflarePurger::class];
 
     /**
      * @var bool
@@ -84,7 +80,7 @@ class SettingsModel extends Model
     /**
      * @var int
      */
-    public $concurrency = 5;
+    public $concurrency = 3;
 
     /**
      * @var string